# Improve last path exposure with better unions