### Identify the Bug or Feature request

Second part of #4506

### Description of the Change

When `FogUtil.exposeLastPath()` needs to union up the results of each step, it used to immediately add the new area to the resulting visible area and token's exposed area. This is really inefficient for a few reasons:
1. It's effectively calculating the same union twice - once for the global exposed area and once for the token exposed area.
2. Each union gets more and more expensive since the geometry gets more complex the further in we get.
3. When the token exposed area is already complex, this makes (2) bad even if the new areas are all simple.

The new approach is to collect the various areas into a list, union them once at the end to produce `visionArea`, then add that to the token exposed area. The union is done by a new method `GeometryUtil.destructiveUnion()` which does a bulk union of area. It avoids a lot of expense by preferring pairwise unions of simple geometries, with only a few unions of complex geometries needed.

As an example of the improvement, the test case in #4506 went from ~30s to ~15s, with the modified part of `FogUtil.exposeLasthPath()` going from ~15s to ~300ms.

### Possible Drawbacks

Should be none.

### Documentation Notes

N/A

### Release Notes

- Improve Expose > Last Path performance with complicated VBL, long paths, or complicated exposed fog.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/RPTools/maptool/4630)
<!-- Reviewable:end -->